Move dataStorage into its own directory
